Correct chord forming
Your major system was creating augmented chords, and your minor system was creating major chords. As such, I corrected the step count. This should also be paired with a fork I'm putting on Instruments/listeners/InventoryClick.java which will make it so the bool minor actually activates the minor functionality instead of the major functionality.